[scripts] Change how some of the testsuite logging works
Unfortunately some of the logging output that was feeding into
testsuite comparisons seems to be too dependent on the versions
of some of the LaTeX packages in use.  This was causing the
test suite to report failures when in fact the latex-mk script
really was working correctly.

I don't see a good way to reliably handle this without just reducing
what gets compared in the log files so do that.  Now any lines
beginning with

@debug:

are ignored for the purposes of pass/fail decisions.  They are in
the logs though to help provide context for other things that are
checked.