[testsuite] Ensure all log lines start with a space for robustness
Different versions of make may or may not include some leading whitespace.
For example one version may produce the following for `make -n` output:

```
    for x in ... ; do
        body
    done
```

While others may have
```
    for x in ... ; do
        body
done
```

and `diff -b` will consider "done" and " done" to be different.  However
" done" and "       done" would be a match as the -b treats one or more
space characters the same.

So in light of this, add in a workaround that ensures all lines start
with a space.