[testsuite] Correctly set the exec/noexec flag for tests specified on the command line
A test name starting with * indicates that make should *not* use -n.  However when
running a single or set of tests from the commandline by listing their names, we don't
usually want to include that *.  That's part of the test, not part of the name.
Handle this case by looking it up.