Fix (hopefully) a bug in how MAKEFLAGS get passed to sub-makes
For reasons I don't recall, I was sticking an extra `-` at the
beginning of each MAKEFLAG when descending into subdirectories with
GNU make.  In the test suite we need to use `--no-print-directory`
to avoid differences in default make behavior with printing the
"Entering/Leaving directory" messages.  But this was getting turned into
`---no-print-directory` (3 dashes) and make fails.

This is why better commit messages and code comments are useful :(