[testsuite] Handle another make implementation variation
The version of bmake encountered on the macos build produced

make: stopped making "default" in /Users/runner/work/latex-mk/latex-mk/testsuite/run/dir1

and this needs to be normalied to

make: stopped in testsuite/run/dir1

to be match other possible variants.  This fixes both the regex for all
non-whitespace characters (used in shortening the path) on non-GNU sed as
well as dropping the
```
making "default"
```
part of the message.

Should fix test subdirs4 on the macos runner.